fixed an issue that caused main window to have build errors when other projects in the solution were added as project references. As a result, the GUI is in a new project
