+++ lunch_project/MainWindow.xaml
@@ -5,91 +5,8 @@
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:lunch_project"
         mc:Ignorable="d"
-        Title="Lunch Transactions" Height="550" Width="850"
-        ResizeMode="NoResize"
-        Background="#FF3F3F46"
-        Loaded="Window_Loaded"
-        VerticalAlignment="Stretch">
-
-    <Grid Margin="0,0,0,-6">
-
-        <DataGrid HorizontalAlignment="Right" Height="494" VerticalAlignment="Top" Width="596"
-                  RowDetailsVisibilityMode="VisibleWhenSelected"
-                  x:Name="dataGridTransactions"
-                  AlternatingRowBackground="LightGray"
-                  CanUserAddRows="False" IsReadOnly="True"
-                  AutoGenerateColumns="False"
-                  CanUserDeleteRows="False"
-                  CanUserResizeRows="False"
-                  CanUserSortColumns="False" ClipboardCopyMode="None"
-                  ScrollViewer.VerticalScrollBarVisibility="Auto"
-                  SelectionMode="Single" Margin="0,36,0,0">
-
-            <DataGrid.RowStyle>
-                <Style TargetType="DataGridRow">
-                    <Setter Property="IsHitTestVisible" Value="False"/>
-                </Style>
-            </DataGrid.RowStyle>
-            <DataGrid.Columns>
-                <DataGridTextColumn Binding="{Binding Path=StudentName}"
-                                    Header="Student" 
-                                    Width="*"
-                                    IsReadOnly="True"
-                                    x:Name="dgcolStudentName"/>
-                <DataGridTextColumn Binding="{Binding Path=Cost, StringFormat=${0}}" 
-                                    Header="Cost"
-                                    Width="75"
-                                    IsReadOnly="True"/>
-                <DataGridTextColumn Binding="{Binding Path=FoodName}" 
-                                    Header="Item" 
-                                    Width="*"
-                                    IsReadOnly="True"/>
-                <DataGridTextColumn Binding="{Binding Path=Time}" 
-                                    Header="Date and Time"
-                                    Width="150"
-                                    IsReadOnly="True"/>
-            </DataGrid.Columns>
-        </DataGrid>
-
-        <TextBox x:Name="txtEnterStudentNum" HorizontalAlignment="Left" Margin="68,12,0,0" 
-                 Text="" TextWrapping="Wrap" VerticalAlignment="Top" Width="120"
-                 KeyDown="OnEnterKeyDownHandler_StudentNum"/>
-
-
-        <DataGrid x:Name="dataGridFoodItems" Margin="10,143,608,211"
-                  CanUserAddRows="False" IsReadOnly="True"
-                  AutoGenerateColumns="False" RowDetailsVisibilityMode="VisibleWhenSelected"
-                  CanUserDeleteRows="False"
-                  CanUserResizeRows="False"
-                  CanUserSortColumns="False"
-                  ScrollViewer.VerticalScrollBarVisibility="Auto"
-                  SelectionMode="Single"
-                  ClipboardCopyMode="None"
-                  SelectionChanged="SelectionChanged_dataGridFoodItems"
-                  GotKeyboardFocus="GotKeyboardFocus_dataGridFoodItems" RowBackground="White">
-
-            <DataGrid.Columns>
-                <DataGridTextColumn Header="Select Item" Binding="{Binding Path=Name}" Width="*" CanUserResize="False"/>
-            </DataGrid.Columns>
-
-        </DataGrid>
-
-
-        <Label Content="Student ID:" HorizontalAlignment="Left" Height="26" Margin="5,8,0,0" VerticalAlignment="Top" Width="79" 
-               FontSize="11" Foreground="White"/>
-        <Label Content="Medical Information:" HorizontalAlignment="Left" Height="28" Margin="5,32,0,0" VerticalAlignment="Top" Width="134" Foreground="Coral" FontWeight="Bold" RenderTransformOrigin="0.515,0.573"/>
-        <TextBlock HorizontalAlignment="Left" Height="77" Margin="11,48,0,0" Text="" TextWrapping="Wrap" VerticalAlignment="Top" Width="198" 
-                   Foreground="Coral" FontSize="11" 
-                   x:Name="txtMedicalInfo"/>
-        <Label HorizontalAlignment="Left" Margin="5,346,0,0" Content="Selected Item:" VerticalAlignment="Top" Foreground="LightGray"/>
-        <TextBlock x:Name="txtSelectedFoodName" HorizontalAlignment="Left" Height="17" Margin="90,351,0,0" Text="" TextWrapping="Wrap" VerticalAlignment="Top" Width="152" Foreground="White" FontWeight="Bold"/>
-        <Label HorizontalAlignment="Left" Margin="5,364,0,0" Content="Cost:" VerticalAlignment="Top" Foreground="LightGray"/>
-        <TextBlock x:Name="txtSelectedFoodCost" HorizontalAlignment="Left" Height="17" Margin="42,369,0,0" Text="" TextWrapping="Wrap" VerticalAlignment="Top" Width="200" Foreground="White" FontWeight="Bold"/>
-        <Label HorizontalAlignment="Left" Margin="5,382,0,0" Content="Description:" VerticalAlignment="Top" Foreground="LightGray"/>
-        <TextBlock x:Name="txtSelectedFoodDescription" HorizontalAlignment="Left" Height="132" Margin="78,388,0,0" Text="" TextWrapping="Wrap" VerticalAlignment="Top" Width="164" Foreground="White" FontSize="11"
-                   ScrollViewer.VerticalScrollBarVisibility="Auto"/>
-        <Button x:Name="buttonUndoTransaction" Content="Undo last transaction" HorizontalAlignment="Left" Margin="400,10,0,0" VerticalAlignment="Top" Width="120" Click="Click_buttonUndoTransaction"/>
-        <Button x:Name="buttonSync" Content="Sync" HorizontalAlignment="Left" Margin="577,10,0,0" VerticalAlignment="Top" Click="Click_buttonSync"/>
-
+        Title="MainWindow" Height="450" Width="800">
+    <Grid>
+        
     </Grid>
 </Window>
