+++ lunch_project/lunch_project.sln
@@ -3,14 +3,10 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31313.79
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CardScannerGUI", "CardScannerGUI.csproj", "{6C558096-6643-48CF-A374-E7BD434C38D5}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{DD0303CA-769F-4E7C-8C9F-626C47658FBA}"
 	ProjectSection(SolutionItems) = preProject
 		..\.gitignore = ..\.gitignore
 	EndProjectSection
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "CardScannerGUI", "CardScannerGUI", "{7951BF4F-1DB7-4234-BC0E-244F586B244C}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Data", "Data", "{DA994DD3-D568-4833-B8A2-B4764703E3D5}"
 EndProject
@@ -20,7 +16,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LunchAPI", "..\LunchAPI\Lun
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Repositories", "Repositories", "{8F64F1DE-CEEE-4BB2-9B33-229EBB9A2913}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Repositories", "..\Repositories\Repositories.csproj", "{3A6F7599-E841-4806-BF7B-84AD5DAA3708}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Repositories", "..\Repositories\Repositories.csproj", "{3A6F7599-E841-4806-BF7B-84AD5DAA3708}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "CardScannerUI", "CardScannerUI", "{29D5C433-3C49-41C9-8D48-5167AE9CD252}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CardScannerUI", "..\CardScannerUI\CardScannerUI.csproj", "{1581AD04-63F7-46D6-AA45-CAA6C2E2DCEA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -28,10 +28,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6C558096-6643-48CF-A374-E7BD434C38D5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6C558096-6643-48CF-A374-E7BD434C38D5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6C558096-6643-48CF-A374-E7BD434C38D5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6C558096-6643-48CF-A374-E7BD434C38D5}.Release|Any CPU.Build.0 = Release|Any CPU
 		{DA1F0DEF-66AA-48D7-A877-393ED03F27BC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{DA1F0DEF-66AA-48D7-A877-393ED03F27BC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DA1F0DEF-66AA-48D7-A877-393ED03F27BC}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -44,14 +40,18 @@ Global
 		{3A6F7599-E841-4806-BF7B-84AD5DAA3708}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3A6F7599-E841-4806-BF7B-84AD5DAA3708}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3A6F7599-E841-4806-BF7B-84AD5DAA3708}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1581AD04-63F7-46D6-AA45-CAA6C2E2DCEA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1581AD04-63F7-46D6-AA45-CAA6C2E2DCEA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1581AD04-63F7-46D6-AA45-CAA6C2E2DCEA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1581AD04-63F7-46D6-AA45-CAA6C2E2DCEA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{6C558096-6643-48CF-A374-E7BD434C38D5} = {7951BF4F-1DB7-4234-BC0E-244F586B244C}
 		{DA1F0DEF-66AA-48D7-A877-393ED03F27BC} = {DA994DD3-D568-4833-B8A2-B4764703E3D5}
 		{3A6F7599-E841-4806-BF7B-84AD5DAA3708} = {8F64F1DE-CEEE-4BB2-9B33-229EBB9A2913}
+		{1581AD04-63F7-46D6-AA45-CAA6C2E2DCEA} = {29D5C433-3C49-41C9-8D48-5167AE9CD252}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {7F1B6B42-4EBC-4EE9-AE89-A76DD1E92206}
